chore: split tests, rename sub directories
